chore: moved other collection / biobank mutations and actions to their new homes
